+++ python/feldera-api-client/feldera_api_client/models/program_status_type_3.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class ProgramStatusType3(str, Enum):
-    COMPILINGRUST = "CompilingRust"
-
-    def __str__(self) -> str:
-        return str(self.value)